app/requirements/requirements.txt
click==8.1.3
colorama==0.4.5
debugpy==1.6.3
dnspython==2.2.1
flake8==4.0.1
Flask==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
numpy==1.23.1
opencv-python==4.6.0.66
Pillow==9.2.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
pymongo==4.1.1
python-decouple==3.6
qrcode==7.3.1
Werkzeug==2.1.2
